feat: Add logout button and finalize MySQL integration
- Add logout button (🚪) to header controls
- Add JavaScript logout() function with confirmation dialog
- Fix SESSION_COOKIE_SECURE to use boolean instead of string
- Update frontend to use /api/transactions instead of /csv endpoint
- Complete MySQL integration with proper session handling

+++ db_mysql.py
@@ -110,7 +110,7 @@ class MySQLReceiptDatabase:
                     business_type VARCHAR(255),
                     notes TEXT,
                     ai_note TEXT,
-                    ai_confidence DECIMAL(5,4),
+                    ai_confidence FLOAT,
                     ai_receipt_merchant VARCHAR(255),
                     ai_receipt_date DATE,
                     ai_receipt_total DECIMAL(10,2),
